Moved custom dialog to components map
