+++ apps/web/src/pages/projects/[project]/widgets/index.tsx
@@ -1,4 +1,4 @@
-import { CreateWidgetDialog } from "@/components/ui/dailog-widget-create";
+import { CreateWidgetDialog } from "@/components/dialog-widget-create";
 import { PageLayout } from "@/components/ui/page-layout";
 import { ListHeading, Paragraph } from "@/components/ui/typography";
 import { useWidgetsHook } from "@/hooks/useWidgetsHook";
